Working on adding an editor view.
Added librarian so that users can add an API token rather than an auth token;
added the ModulusView class which has the events required to start, stop, and
restart a project. The ModulusView class propmts for the name of the project
to perform an action on, then (not complete) performs the action in the
specified project.